Частично поправил модель. Переделал запрос на авторизацю

+++ Models/Role.cs
@@ -1,6 +1,6 @@
 namespace circle_competitions_monitoring.Models {
     public class Role {
-        public int id { get; set; }
+        public byte id { get; set; }
         public string name { get; set; }
     }
 }
